[iOS/#329] 일간 학습 통계 로직 구현 (#340)
* fix: label 색상 수정

* feat: VC 연결 by Coordinator pattern

* fix: SwiftUIView padding 적용

* feat: CustomDatePicker 추가 (다크모드 대응)

* fix: View Cropping 수정

* feat: SwiftUIView Mock Data 제거

* refactor: Boiler Plate 삭제

* refactor: Logger 추가 및 중복 코드 제거

* feat: View, View, ViewModel 삼자 바인딩

* feat: DailyChartView Logic 완성

* feat: 기타 카테고리 계산 로직 추가

* feat: 차트 페이지 SE 등 Frame 작은 Device 대응

* feat: 주간 학습 네트워크 구성 및 ViewModel

* feat: 주간 학습 통계 세팅

* fix: 일부 오류 수정

- 처음에 오늘 학습 기록이 로드되지 않는 오류
- 탭 이동 후 돌아왔을 때 오늘 날짜로 변경되게 수정

+++ iOS/FlipMate/FlipMate/Data/Network/DataMapping/ChartScene/DailyChartLogResponseDTO.swift
@@ -10,7 +10,7 @@ import Foundation
 struct DailyChartLogResponseDTO: Decodable {
     let todayTime: Int
     let categories: [CategoryDTO]?
-    let percentage: Int
+    let percentage: Double
     
     private enum CodingKeys: String, CodingKey {
         case todayTime = "total_time"
